move TypeBeanInstanceProducer as per the todo
