+++ hibernate-core/src/main/java/org/hibernate/boot/spi/BootstrapContext.java
@@ -14,6 +14,7 @@ import org.hibernate.boot.archive.scan.spi.ScanEnvironment;
 import org.hibernate.boot.archive.scan.spi.ScanOptions;
 import org.hibernate.boot.archive.spi.ArchiveDescriptorFactory;
 import org.hibernate.boot.internal.ClassmateContext;
+import org.hibernate.boot.internal.TypeBeanInstanceProducer;
 import org.hibernate.boot.model.convert.spi.ConverterDescriptor;
 import org.hibernate.boot.model.relational.AuxiliaryDatabaseObject;
 import org.hibernate.boot.registry.StandardServiceRegistry;
@@ -61,7 +62,7 @@ public interface BootstrapContext {
 	/**
 	 * The {@link BeanInstanceProducer} to use when creating custom type references.
 	 *
-	 * @implNote Usually a {@link org.hibernate.boot.model.TypeBeanInstanceProducer}.
+	 * @implNote Usually a {@link TypeBeanInstanceProducer}.
 	 */
 	BeanInstanceProducer getCustomTypeProducer();
